Move QueryColumn interface into its own class
