+++ frontend/app/components/wp-fast-table/wp-table-group-by.ts
@@ -27,12 +27,10 @@
 // ++
 
 import {QueryGroupByResource} from '../api/api-v3/hal-resources/query-group-by-resource.service';
-import {
-  QueryResource,
-  QueryColumn
-} from '../api/api-v3/hal-resources/query-resource.service';
+import {QueryResource} from '../api/api-v3/hal-resources/query-resource.service';
 import {QuerySchemaResourceInterface} from '../api/api-v3/hal-resources/query-schema-resource.service';
-import {WorkPackageTableBaseState} from "./wp-table-base";
+import {WorkPackageTableBaseState} from './wp-table-base';
+import {QueryColumn} from '../wp-query/query-column';
 
 export class WorkPackageTableGroupBy extends WorkPackageTableBaseState<QueryGroupByResource | undefined> {
   public available:QueryGroupByResource[] = [];
