+++ frontend/app/components/wp-fast-table/wp-table-sort-by.ts
@@ -31,9 +31,10 @@ import {
   QUERY_SORT_BY_DESC,
   QuerySortByResource
 } from '../api/api-v3/hal-resources/query-sort-by-resource.service';
-import {QueryColumn, QueryResource} from '../api/api-v3/hal-resources/query-resource.service';
+import {QueryResource} from '../api/api-v3/hal-resources/query-resource.service';
 import {QuerySchemaResourceInterface} from '../api/api-v3/hal-resources/query-schema-resource.service';
 import {WorkPackageTableBaseState} from './wp-table-base';
+import {QueryColumn} from '../wp-query/query-column';
 
 export class WorkPackageTableSortBy extends WorkPackageTableBaseState<QuerySortByResource[]> {
   public available: QuerySortByResource[] = [];
