+++ frontend/app/components/wp-inline-create/inline-create-row-builder.ts
@@ -1,4 +1,3 @@
-import {WorkPackageTableRow} from '../wp-fast-table/wp-table.interfaces';
 import {TableRowEditContext} from '../wp-edit-form/table-row-edit-context';
 import {WorkPackageEditForm} from '../wp-edit-form/work-package-edit-form';
 import {injectorBridge} from '../angular/angular-injector-bridge.functions';
@@ -6,22 +5,18 @@ import {
   WorkPackageResource,
   WorkPackageResourceInterface
 } from '../api/api-v3/hal-resources/work-package-resource.service';
-import {checkedClassName} from '../wp-fast-table/builders/ui-state-link-builder';
 import {rowId} from '../wp-fast-table/helpers/wp-table-row-helpers';
 import {States} from '../states.service';
 import {WorkPackageTableSelection} from '../wp-fast-table/state/wp-table-selection.service';
 import {WorkPackageTableColumnsService} from '../wp-fast-table/state/wp-table-columns.service';
-import {CellBuilder} from '../wp-fast-table/builders/cell-builder';
 import {
   internalDetailsColumn,
-  rowClassName,
-  SingleRowBuilder
+  rowClassName
 } from '../wp-fast-table/builders/rows/single-row-builder';
+import {WorkPackageTable} from '../wp-fast-table/wp-fast-table';
+import {RowRefreshBuilder} from '../wp-fast-table/builders/rows/row-refresh-builder';
 import IScope = angular.IScope;
-import {scopeDestroyed$} from "../../helpers/angular-rx-utils";
-import {WorkPackageTable} from "../wp-fast-table/wp-fast-table";
-import {QueryColumn} from '../api/api-v3/hal-resources/query-resource.service';
-import {RowRefreshBuilder} from "../wp-fast-table/builders/rows/row-refresh-builder";
+import {QueryColumn} from '../wp-query/query-column';
 
 export const inlineCreateRowClassName = 'wp-inline-create-row';
 export const inlineCreateCancelClassName = 'wp-table--cancel-create-link';
